Use applicationDirPath() rather than currentPath()
-When building a relative path to supplementary executables (e.g. extended
mode toggler), we were using QDir::currentPath(), which is "Returns the
absolute path of the application's current directory."  It turns out that
they meant working directory.  When the CPI is launched from the final
page of the NSIS installer, the working directory is not correct, and thus
the relative path to executables is wrong.
-Instead, we're now using QCoreApplication::applicationDirPath(), which
actually just provides the path to the directory containing the CPI
executable itself.  This is more reliable, unless you move files around
inside of Program Files, at which point it's out of our hands.